Refactor code to move login route from AuthController to LoginController
