+++ public/views/logout.pug
@@ -13,3 +13,4 @@ html(lang="en")
           div#title Logout
       #content
         h2='You\'ve been logged out from ' + serviceName
+        a(href="https://tko-aly.fi") Return to TKO-äly website
